+++ lib/PayPal/Api/CreditCard.php
@@ -12,6 +12,10 @@ class CreditCard extends Resource implements IResource {
 
 	private static $credential;
 	
+	/**
+	 *
+	 * @deprected. Pass ApiContext to create/get methods instead
+	 */
 	public static function setCredential($credential) {
 		self::$credential = $credential;
 	}
@@ -26,6 +30,7 @@ class CreditCard extends Resource implements IResource {
 
 	/**
 	 * Getter for id
+	 * @return string
 	 */ 
 	public function getId() {
 		return $this->id;
@@ -41,6 +46,7 @@ class CreditCard extends Resource implements IResource {
 
 	/**
 	 * Getter for valid_until
+	 * @return string
 	 */ 
 	public function getValid_until() {
 		return $this->valid_until;
@@ -56,6 +62,7 @@ class CreditCard extends Resource implements IResource {
 
 	/**
 	 * Getter for state
+	 * @return string
 	 */ 
 	public function getState() {
 		return $this->state;
@@ -71,6 +78,7 @@ class CreditCard extends Resource implements IResource {
 
 	/**
 	 * Getter for payer_id
+	 * @return string
 	 */ 
 	public function getPayer_id() {
 		return $this->payer_id;
@@ -86,6 +94,7 @@ class CreditCard extends Resource implements IResource {
 
 	/**
 	 * Getter for type
+	 * @return string
 	 */ 
 	public function getType() {
 		return $this->type;
@@ -101,6 +110,7 @@ class CreditCard extends Resource implements IResource {
 
 	/**
 	 * Getter for number
+	 * @return string
 	 */ 
 	public function getNumber() {
 		return $this->number;
@@ -116,6 +126,7 @@ class CreditCard extends Resource implements IResource {
 
 	/**
 	 * Getter for expire_month
+	 * @return string
 	 */ 
 	public function getExpire_month() {
 		return $this->expire_month;
@@ -131,6 +142,7 @@ class CreditCard extends Resource implements IResource {
 
 	/**
 	 * Getter for expire_year
+	 * @return string
 	 */ 
 	public function getExpire_year() {
 		return $this->expire_year;
@@ -146,6 +158,7 @@ class CreditCard extends Resource implements IResource {
 
 	/**
 	 * Getter for cvv2
+	 * @return string
 	 */ 
 	public function getCvv2() {
 		return $this->cvv2;
@@ -161,6 +174,7 @@ class CreditCard extends Resource implements IResource {
 
 	/**
 	 * Getter for first_name
+	 * @return string
 	 */ 
 	public function getFirst_name() {
 		return $this->first_name;
@@ -176,6 +190,7 @@ class CreditCard extends Resource implements IResource {
 
 	/**
 	 * Getter for last_name
+	 * @return string
 	 */ 
 	public function getLast_name() {
 		return $this->last_name;
@@ -191,6 +206,7 @@ class CreditCard extends Resource implements IResource {
 
 	/**
 	 * Getter for billing_address
+	 * @return PayPal\Api\Address
 	 */ 
 	public function getBilling_address() {
 		return $this->billing_address;
@@ -206,6 +222,7 @@ class CreditCard extends Resource implements IResource {
 
 	/**
 	 * Getter for links
+	 * @return PayPal\Api\Link
 	 */ 
 	public function getLinks() {
 		return $this->links;
@@ -217,15 +234,17 @@ class CreditCard extends Resource implements IResource {
 	 * @path /v1/vault/credit-card
 	 * @method POST
 	  
-	 * @param PayPal\Rest\ApiContext $apiContext optional	  	 
+	 * @param PayPal\Rest\ApiContext $apiContext optional
 	 */
 	public function create( $apiContext=null) {
 		$payLoad = $this->toJSON();	
 		if($apiContext == null) {
 			$apiContext = new ApiContext(self::$credential);
 		}
-		$call = new Call();		
-		$json = $call->execute("/v1/vault/credit-card", "POST", $payLoad, $apiContext);
+		$call = new \PPRestCall($apiContext);		
+		$json = $call->execute( array('PayPal\Rest\RestHandler'),
+			"/v1/vault/credit-card", 
+			"POST", $payLoad);
 		$this->fromJson($json);
  		return $this; 		
  	}
@@ -233,16 +252,21 @@ class CreditCard extends Resource implements IResource {
 	/**
 	 * @path /v1/vault/credit-card/:credit-card-id
 	 * @method GET
-	 * @param string $creditcardid	  	 
+	 * @param string $creditcardid	  
+	 * @param PayPal\Rest\ApiContext $apiContext optional
 	 */
-	public static function get( $creditcardid) {
+	public static function get( $creditcardid, $apiContext=null) {
 		if (($creditcardid == null) || (strlen($creditcardid) <= 0)) {
 			throw new \InvalidArgumentException("creditcardid cannot be null or empty");
 		}
 		$payLoad = "";
-		
-		$apiContext = new ApiContext(self::$credential);		$call = new Call();		
-		$json = $call->execute("/v1/vault/credit-card/$creditcardid", "GET", $payLoad, $apiContext);
+		if($apiContext == null) {
+			$apiContext = new ApiContext(self::$credential);
+		}
+		$call = new \PPRestCall($apiContext);		
+		$json = $call->execute( array('PayPal\Rest\RestHandler'),
+			"/v1/vault/credit-card/$creditcardid", 
+			"GET", $payLoad);
 		$ret = new CreditCard();
 		$ret->fromJson($json);
 		return $ret;