+++ lib/PayPal/Api/Payment.php
@@ -12,6 +12,10 @@ class Payment extends Resource implements IResource {
 
 	private static $credential;
 	
+	/**
+	 *
+	 * @deprected. Pass ApiContext to create/get methods instead
+	 */
 	public static function setCredential($credential) {
 		self::$credential = $credential;
 	}
@@ -26,6 +30,7 @@ class Payment extends Resource implements IResource {
 
 	/**
 	 * Getter for id
+	 * @return string
 	 */ 
 	public function getId() {
 		return $this->id;
@@ -41,6 +46,7 @@ class Payment extends Resource implements IResource {
 
 	/**
 	 * Getter for create_time
+	 * @return string
 	 */ 
 	public function getCreate_time() {
 		return $this->create_time;
@@ -56,6 +62,7 @@ class Payment extends Resource implements IResource {
 
 	/**
 	 * Getter for update_time
+	 * @return string
 	 */ 
 	public function getUpdate_time() {
 		return $this->update_time;
@@ -71,6 +78,7 @@ class Payment extends Resource implements IResource {
 
 	/**
 	 * Getter for state
+	 * @return string
 	 */ 
 	public function getState() {
 		return $this->state;
@@ -86,6 +94,7 @@ class Payment extends Resource implements IResource {
 
 	/**
 	 * Getter for intent
+	 * @return string
 	 */ 
 	public function getIntent() {
 		return $this->intent;
@@ -101,6 +110,7 @@ class Payment extends Resource implements IResource {
 
 	/**
 	 * Getter for payer
+	 * @return PayPal\Api\Payer
 	 */ 
 	public function getPayer() {
 		return $this->payer;
@@ -116,6 +126,7 @@ class Payment extends Resource implements IResource {
 
 	/**
 	 * Getter for transactions
+	 * @return PayPal\Api\Transaction
 	 */ 
 	public function getTransactions() {
 		return $this->transactions;
@@ -131,6 +142,7 @@ class Payment extends Resource implements IResource {
 
 	/**
 	 * Getter for redirect_urls
+	 * @return PayPal\Api\RedirectUrls
 	 */ 
 	public function getRedirect_urls() {
 		return $this->redirect_urls;
@@ -146,6 +158,7 @@ class Payment extends Resource implements IResource {
 
 	/**
 	 * Getter for links
+	 * @return PayPal\Api\Link
 	 */ 
 	public function getLinks() {
 		return $this->links;
@@ -167,14 +180,19 @@ class Payment extends Resource implements IResource {
 	 *			payee_id,
 	 *			sort_by,
 	 *			sort_order,
-	 *			All other keys in the map are ignored by the SDK	  	 
+	 *			All other keys in the map are ignored by the SDK	  
+	 * @param PayPal\Rest\ApiContext $apiContext optional
 	 */
-	public static function all($params) {
+	public static function all($params, $apiContext=null) {
 		$payLoad = "";
 		$allowedParams = array('count' => 1, 'start_id' => 1, 'start_index' => 1, 'start_time' => 1, 'end_time' => 1, 'payee_id' => 1, 'sort_by' => 1, 'sort_order' => 1, );		
-		
-		$apiContext = new ApiContext(self::$credential);		$call = new Call();		
-		$json = $call->execute("/v1/payments/payment?" . http_build_query(array_intersect_key($params, $allowedParams)), "GET", $payLoad, $apiContext);
+		if($apiContext == null) {
+			$apiContext = new ApiContext(self::$credential);
+		}
+		$call = new \PPRestCall($apiContext);		
+		$json = $call->execute( array('PayPal\Rest\RestHandler'),
+			"/v1/payments/payment?" . http_build_query(array_intersect_key($params, $allowedParams)), 
+			"GET", $payLoad);
 		$ret = new PaymentHistory();
 		$ret->fromJson($json);
 		return $ret;
@@ -185,15 +203,17 @@ class Payment extends Resource implements IResource {
 	 * @path /v1/payments/payment
 	 * @method POST
 	  
-	 * @param PayPal\Rest\ApiContext $apiContext optional	  	 
+	 * @param PayPal\Rest\ApiContext $apiContext optional
 	 */
 	public function create( $apiContext=null) {
 		$payLoad = $this->toJSON();	
 		if($apiContext == null) {
 			$apiContext = new ApiContext(self::$credential);
 		}
-		$call = new Call();		
-		$json = $call->execute("/v1/payments/payment", "POST", $payLoad, $apiContext);
+		$call = new \PPRestCall($apiContext);		
+		$json = $call->execute( array('PayPal\Rest\RestHandler'),
+			"/v1/payments/payment", 
+			"POST", $payLoad);
 		$this->fromJson($json);
  		return $this; 		
  	}
@@ -201,16 +221,21 @@ class Payment extends Resource implements IResource {
 	/**
 	 * @path /v1/payments/payment/:payment-id
 	 * @method GET
-	 * @param string $paymentid	  	 
+	 * @param string $paymentid	  
+	 * @param PayPal\Rest\ApiContext $apiContext optional
 	 */
-	public static function get( $paymentid) {
+	public static function get( $paymentid, $apiContext=null) {
 		if (($paymentid == null) || (strlen($paymentid) <= 0)) {
 			throw new \InvalidArgumentException("paymentid cannot be null or empty");
 		}
 		$payLoad = "";
-		
-		$apiContext = new ApiContext(self::$credential);		$call = new Call();		
-		$json = $call->execute("/v1/payments/payment/$paymentid", "GET", $payLoad, $apiContext);
+		if($apiContext == null) {
+			$apiContext = new ApiContext(self::$credential);
+		}
+		$call = new \PPRestCall($apiContext);		
+		$json = $call->execute( array('PayPal\Rest\RestHandler'),
+			"/v1/payments/payment/$paymentid", 
+			"GET", $payLoad);
 		$ret = new Payment();
 		$ret->fromJson($json);
 		return $ret;
@@ -221,7 +246,7 @@ class Payment extends Resource implements IResource {
 	 * @path /v1/payments/payment/:payment-id/execute
 	 * @method POST
 	 * @param PaymentExecution $payment_execution	  
-	 * @param PayPal\Rest\ApiContext $apiContext optional	  	 
+	 * @param PayPal\Rest\ApiContext $apiContext optional
 	 */
 	public function execute( $payment_execution, $apiContext=null) {
 		if ($payment_execution == null) {
@@ -234,8 +259,10 @@ class Payment extends Resource implements IResource {
 		if($apiContext == null) {
 			$apiContext = new ApiContext(self::$credential);
 		}
-		$call = new Call();		
-		$json = $call->execute("/v1/payments/payment/{$this->getId()}/execute", "POST", $payLoad, $apiContext);
+		$call = new \PPRestCall($apiContext);		
+		$json = $call->execute( array('PayPal\Rest\RestHandler'),
+			"/v1/payments/payment/{$this->getId()}/execute", 
+			"POST", $payLoad);
 		$this->fromJson($json);
  		return $this; 		
  	}