+++ lib/PayPal/Api/Resource.php
@@ -1,12 +1,10 @@
 <?php 
 
 namespace PayPal\Api;
-use PayPal\Common\Model;
-
 /**
  * 
  */
-class Resource extends Model {
+class Resource extends \PPModel {