+++ lib/PayPal/Rest/ApiContext.php
@@ -6,7 +6,7 @@ namespace PayPal\Rest;
  * Call level parameters such as
  * request id, credentials etc
  */
-class ApiContext {
+class ApiContext extends \PPApiContext {
 	
 	/**
 	 * OAuth Credentials to use for this call
@@ -22,7 +22,6 @@ class ApiContext {
 	 */
 	private $requestId;
 	
-	
 	/**
 	 * 
 	 */
@@ -32,10 +31,11 @@ class ApiContext {
 	
 	public function getrequestId() {
 		if($this->requestId == null) {
-			$this->requestId = $this->generaterequestId();
+			$this->requestId = $this->generateRequestId();
 		}
 		return $this->requestId;
 	}
+	
 	
 	/**
 	 *