+++ sample/payments/ListPayments.php
@@ -12,12 +12,6 @@
 require __DIR__ . '/../bootstrap.php';
 use PayPal\Api\Payment;
 
-// ### Authentication
-// Pass in a `OAuthTokenCredential` object
-// explicilty to authenticate the call. 
-// If you skip this step, the client id/secret
-// set in the config file will be used. 
-Payment::setCredential($cred);
 
 // ### Retrieve payment
 // Retrieve the PaymentHistory object by calling the
@@ -25,8 +19,9 @@ Payment::setCredential($cred);
 // and pass a Map object that contains
 // query parameters for paginations and filtering.
 // Refer the method doc for valid values for keys
+// (See bootstrap.php for more on `ApiContext`)
 try {
-	$payments = Payment::all(array('count' => 10, 'start_index' => 5));	
+	$payments = Payment::all(array('count' => 10, 'start_index' => 5), $apiContext);	
 } catch (\PPConnectionException $ex) {
 	echo "Exception:" . $ex->getMessage() . PHP_EOL;
 	var_dump($ex->getData());