+++ sample/sale/GetSale.php
@@ -9,17 +9,12 @@ require __DIR__ . '/../bootstrap.php';
 use PayPal\Api\Sale;
 
 $saleId = '3RM92092UW5126232';
-// ### Authentication
-// Pass in a `OAuthTokenCredential` object
-// explicilty to authenticate the call. 
-// If you skip this step, the client id/secret
-// set in the config file will be used. 
-Sale::setCredential($cred);
+
 try {	
 	// ### Retrieve the sale object
 	// Pass the ID of the sale
 	// transaction from your payment resource.
-	$sale = Sale::get($saleId);
+	$sale = Sale::get($saleId, $apiContext);
 } catch (\PPConnectionException $ex) {
 	echo "Exception:" . $ex->getMessage() . PHP_EOL;
 	var_dump($ex->getData());